Issue #14937: Migrated section 4.8.2.1 One Variable Per Declaration to chapter wise testing
