+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/XdocsPagesTest.java
@@ -239,7 +239,6 @@ public class XdocsPagesTest {
             "4.5.2 Indent continuation lines at least +4 spaces",
             "4.6.1 Vertical Whitespace",
             "4.6.2 Horizontal whitespace",
-            "4.8.2.1 One variable per declaration",
             "4.8.2.2 Declared when needed",
             "4.8.3.2 No C-style array declarations",
             "4.8.4.1 Indentation",
