+++ src/xdocs/google_style.xml
@@ -1143,7 +1143,7 @@
                   </a>
                 </td>
                 <td>
-                  <a href="styleguides/google-java-style-20180523/javaguide.html#s4.8.2.1-variables-per-declaration">
+                  <a href="styleguides/google-java-style-20220203/javaguide.html#s4.8.2.1-variables-per-declaration">
                     4.8.2.1 One variable per declaration</a>
                 </td>
                 <td>
@@ -1157,7 +1157,7 @@
                   <a href="https://github.com/search?q=path%3Asrc%2Fmain%2Fresources%20path%3A**%2Fgoogle_checks.xml+repo%3Acheckstyle%2Fcheckstyle+MultipleVariableDeclarations">
                     config</a>
                   <br />
-                  <a href="https://github.com/checkstyle/checkstyle/blob/master/src/it/java/com/google/checkstyle/test/chapter4formatting/rule4821onevariableperline/MultipleVariableDeclarationsTest.java">
+                  <a href="https://github.com/checkstyle/checkstyle/blob/master/src/it/java/com/google/checkstyle/test/chapter4formatting/rule4821onevariableperline/OneVariablePerDeclarationTest.java">
                     test</a>
                 </td>
               </tr>
